Add initial UI
I don't need each player to manually choose their cell via a prompt. so
i'm removing the getPlayerCell function from the GameFlow module.
The playround function now take two paramaters(row and col) to facilitate
the marksetting on the board.  In the console version, those values were
got in the playround func via the getPlayerCell func, now they will be passed
directly to playround from the UI. We don't need to alert which player turn its
anymore too.
in the chekwinner funnc, i remove the playround() call because we don't need
it again in the UI, the playround(with 2 param) need to be called from the UI
for the 2 param to be passed to it.
we are not calling playround from GameFlow module too, because of the same reason
above.
Add a DisplayGame module to encapsulate all the UI logics..