Add player scores to the display
I delare a variable name score in the player module and return it.
Each player score will be incremented in checkwinner() each time the player wins a round.
I then return the two players in the Gameflow module, to be able to access them in
the display to render the score.